requirements.txt
# Project Enclave - Core Dependencies

# LLM & AI
anthropic>=0.40.0
langgraph>=0.2.0
langchain-core>=0.3.0

# Database
supabase>=2.0.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0

# Configuration
pydantic>=2.5.0
pyyaml>=6.0.0

# Email
jinja2>=3.1.0

# CLI / Dashboard
typer>=0.12.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: Dashboard
# streamlit>=1.35.0

# Optional: Enrichment
# shodan>=1.31.0
